init: start migration to typescript
