+++ .eslintrc.json
@@ -6,27 +6,42 @@
     "jest": true,
     "node": true
   },
+  "parser": "@typescript-eslint/parser",
   "parserOptions": {
     "sourceType": "module",
     "ecmaVersion": 2021,
     "ecmaFeatures": {
       "impliedStrict": true
+    },
+    "parserOptions": {
+      "project": "./tsconfig.json"
     }
   },
   "settings": {
-    "noInlineConfig": true
+    "noInlineConfig": true,
+    "import/resolver": {
+      "node": {
+        "extensions": [".js", ".ts"]
+      }
+    }
   },
   "extends": [
     "eslint:recommended",
-    "plugin:node/recommended",
     "airbnb-base",
-    "prettier"
+    "prettier",
+    "plugin:import/typescript"
+  ],
+  "plugins": [
+    "@typescript-eslint"
   ],
   "rules": {
     "import/extensions": "off",
+    "import/no-unresolved": "off",
     "linebreak-style": "off",
     "node/no-unsupported-features/es-syntax": "off",
     "no-underscore-dangle": "off",
-    "import/prefer-default-export": "off"
+    "import/prefer-default-export": "off",
+    "dot-notation":"off",
+    "no-unused-vars": "warn"
   }
 }
